requirements.txt
python-dotenv
openai
chromadb
langchain
pypdf
tiktoken
unstructured
gitpython
flask